memshare: Free the memory after XPU unlocking is done
Currently memshare driver is listening to notification
SUBSYS_AFTER_SHUTDOWN, which doesn't guarantee the unlocking
of XPU for the clients sitting on peripheral. This could lead
to a XPU violation. Move the notification to free up memory
during SUBSYS_BEFORE_POWERUP to ensure that the allocated memory
XPU has been removed.

Change-Id: Idff7972e78c8e8deb987179b96741effb06799b6
CRs-Fixed: 769787
Signed-off-by: Katish Paran <kparan@codeaurora.org>